+++ src/kernels/xilinx/reducesum4d.cpp
@@ -88,11 +88,11 @@ void ReduceSumRank4Axes012(
         buff_rslt[d3] = buff_rslt[d3] + pow_rslt;
 
         //=====================================================
-        if( d3 == ((unsigned long)dim3-1) ){
+        if( d3 == (dim3-1) ){
             d3=0;
-            if( d2 == ((unsigned long)dim2-1) ){
+            if( d2 == (dim2-1) ){
                 d2=0;
-                if( d1 == ((unsigned long)dim1-1)){
+                if( d1 == (dim1-1)){
                     d1=0;
                     d0++;
                 }else{
